+++ packages/rma-frontend/src/app/sales-ui/services/sales.service.ts
@@ -7,12 +7,14 @@ import {
   Item,
   APIResponse,
   SerialAssign,
+  AggregatedDocument,
 } from '../../common/interfaces/sales.interface';
 import {
   AUTHORIZATION,
   BEARER_TOKEN_PREFIX,
   ACCESS_TOKEN,
   DEFAULT_SELLING_PRICE_LIST,
+  HUNDRED_NUMBERSTRING,
 } from '../../constants/storage';
 import {
   LIST_SALES_INVOICE_ENDPOINT,
@@ -60,7 +62,6 @@ import {
   NON_SERIAL_ITEM,
   TERRITORY,
 } from '../../constants/app-string';
-import { DeliveryNoteItemInterface } from '../view-sales-invoice/serials/serials-datasource';
 import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Injectable({
@@ -628,6 +629,7 @@ export class SalesService {
     const params = new HttpParams({
       fromObject: {
         filters: `[["against_sales_invoice","=","${invoice_name}"],["is_return","=","0"]]`,
+        limit_page_length: (HUNDRED_NUMBERSTRING * 10).toString(),
       },
     });
     return this.getHeaders().pipe(
@@ -775,15 +777,9 @@ export class SalesService {
     return this.http.get<any>(API_INFO_ENDPOINT);
   }
 
-  getAggregatedDocument(
-    res: {
-      items: DeliveryNoteItemInterface[];
-      total?: number;
-      total_qty?: number;
-    }[],
-  ) {
+  getAggregatedDocument(res: AggregatedDocument[]) {
     const itemsHashMap = {};
-    const doc = res[0];
+    const doc: any = res[0];
     doc.total = 0;
     doc.total_qty = 0;
     res.forEach(document => {